+++ bin/runme_python.sh
@@ -22,15 +22,10 @@
 #
 params=$*
 
-
 #
 # Find out our build dir
 #
 . etc/build_dir.sh
-
-export build_root=BUILD-$KERNEL-$ARCH-python
-export BUILD_DIR=${build_root}/${BUILD_TYPE_DIR}
-mkdir -p $BUILD_DIR
 
 
 #
@@ -38,43 +33,42 @@ mkdir -p $BUILD_DIR
 #
 export FLAGS=""
 export CMAKE_FLAGS=""
-unset PYTHONEXE
-unset PYTHON
-unset PYTHONDIR
-unset PYTHON_SITEDIR
-unset PYTHON_VERSION
 
+#
+# Common flags
+#
+export BUILD_PYTHON=ON
 
 #
 # These are some of the expensive mrv2 options
 #
-export BUILD_PYTHON=OFF
-export MRV2_PYFLTK=OFF
+export MRV2_PYFLTK=ON
 export MRV2_PYBIND11=ON
-export MRV2_NETWORK=OFF
-export MRV2_PDF=OFF
-export MRV2_PYTHON=/usr/bin/python3.11
+export MRV2_NETWORK=ON
+export MRV2_PDF=ON
 
 #
 # These are some of the expensive TLRENDER options
 #
-export TLRENDER_AV1=OFF
-export TLRENDER_FFMPEG=OFF
-export TLRENDER_FFMPEG_MINIMAL=OFF
-export TLRENDER_NDI=OFF
-export TLRENDER_NET=OFF
-export TLRENDER_RAW=OFF
-export TLRENDER_SGI=OFF
-export TLRENDER_STB=OFF
-export TLRENDER_TIFF=OFF
+export TLRENDER_AV1=ON
+export TLRENDER_FFMPEG=ON
+export TLRENDER_FFMPEG_MINIMAL=ON
+export TLRENDER_EXR=ON
+export TLRENDER_JPEG=ON
+export TLRENDER_NDI=ON
+export TLRENDER_NET=ON
+export TLRENDER_RAW=ON
+export TLRENDER_SGI=ON
+export TLRENDER_STB=ON
+export TLRENDER_TIFF=ON
 export TLRENDER_USD=OFF
-export TLRENDER_VPX=OFF
+export TLRENDER_VPX=ON
 export TLRENDER_WAYLAND=ON
 export TLRENDER_X11=ON
-export TLRENDER_YASM=OFF
+export TLRENDER_YASM=ON
 
 echo
 echo "Saving compile log to $BUILD_DIR/compile.log ..."
 echo
-cmd="./runme_nolog.sh --build-dir ${build_root} --ask $params 2>&1 | tee $BUILD_DIR/compile.log"
+cmd="./runme_nolog.sh --ask $params 2>&1 | tee $BUILD_DIR/compile.log"
 run_cmd $cmd