+++ mrv2/lib/mrvGL/mrvGLDefines.h
@@ -21,3 +21,10 @@
 // #define USE_GL_CHECKS 1
 
 // #define MRV2_DEBUG_GL 1  // Set this to debug OpenGL errors with a callback
+
+// Set this to debug the video times as they come from the player video
+// timeout.
+// #define DEBUG_VIDEO_CALLBACK 1  
+
+// Set this to debug the video times as they come from the timeline timeout.
+// #define DEBUG_TIMELINE_CALLBACK 1