+++ mrv2/lib/mrvGL/mrvTimelineViewport.cpp
@@ -1002,11 +1002,19 @@ namespace mrv
         const TimelinePlayer* sender) noexcept
     {
         TLRENDER_P();
+        
+#ifdef DEBUG_VIDEO_CALLBACK
+        if (!p.videoData.empty() && !values.empty() &&
+            values[0].time != p.videoData[0].time)
+        {
+            std::cerr << values[0].time << std::endl;
+        }
+#endif
+        
         p.videoData = values;
 
         if (p.resizeWindow)
         {
-            p.switchClip = false;
             if (!p.presentation)
                 resizeWindow();
             else
@@ -1015,9 +1023,14 @@ namespace mrv
         else if (p.frameView && p.switchClip)
         {
             frameView();
-            p.switchClip = false;
         }
 
+        if (p.switchClip)
+        {
+            p.switchClip = false;
+            p.droppedFrames = 0;
+        }
+        
         _getTags();
         int layerId = sender->videoLayer();
         p.missingFrame = false;