+++ mrv2/docs/HISTORY.md
@@ -6,8 +6,6 @@ It is a bug fix release for the regressions in v1.1.4 which was rushed:
 
 - Added a FAQ on the documentation about Linux XWayland and Wayland on
   why it could make playback slow if your OS was not configured properly.
-  I added current instructions on Ubuntu 22.04.4 LTS to get full NVidia
-  support working properly on both XWayland and Wayland (yeah!).
 - This is the first version that supports Wayland and XWayland.
   For instructions on how to set XWayland and Wayland on Ubuntu 22.04.4 LTS
   with NVidia drivers, please see Help->Documentation the FAQ section.
@@ -24,6 +22,9 @@ It is a bug fix release for the regressions in v1.1.4 which was rushed:
   which would crash Wayland.
 - mrv2 can now be built with the pre-installed python, without having to
   build python itself.  Of course, in that case, mrv2 cannot be re-distributed.
+- Added a python translating system for plug-ins.
+- Translated all update-mrv2.py strings into Spanish.
+
 
 
 v1.1.4